Add files via upload
Since most of the code was developed in arduino mega, problems with memory for variables were not detected, however, when trying to switch to arduino pro mini, problems started appearing to fix these issues, a last revision of the code was made. , excluding a series of variables, and changing the type of others, always for a variable that occupies a smaller number of bytes in memory.